Add Tenancy to URI Signing endpoints and refactor riak code into a riaksvc package.
