+++ traffic_ops/traffic_ops_golang/urisigning_test.go
@@ -20,14 +20,8 @@ package main
  */
 
 import (
-	"crypto/tls"
 	"encoding/json"
-	"errors"
 	"testing"
-
-	"github.com/basho/riak-go-client"
-	"github.com/jmoiron/sqlx"
-	"gopkg.in/DATA-DOG/go-sqlmock.v1"
 )
 
 const (
@@ -132,114 +126,6 @@ const (
 `
 )
 
-type MockStorageCluster struct {
-	Bucket  string
-	Key     string
-	Running bool
-}
-
-func (mc MockStorageCluster) Stop() error {
-	if mc.Running == true {
-		mc.Running = false
-	} else {
-		return errors.New("the cluster is not started")
-	}
-	return nil
-}
-
-func (mc MockStorageCluster) Start() error {
-	if mc.Running == false {
-		mc.Running = true
-	} else {
-		return errors.New("the cluster is already started")
-	}
-	return nil
-}
-
-func (mc MockStorageCluster) Execute(command riak.Command) error {
-	return nil
-}
-
-func TestFetchObjectValues(t *testing.T) {
-	cluster := &MockStorageCluster{
-		Running: true,
-	}
-
-	_, err := fetchObjectValues("myobject", "bucket", cluster)
-	if err != nil {
-		t.Error("expected nil error got ", err)
-	}
-
-	_, err = fetchObjectValues("", "bucket", cluster)
-	if err == nil {
-		t.Error("expected an error because key is empty but got no error")
-	}
-
-	_, err = fetchObjectValues("myobject", "", cluster)
-	if err == nil {
-		t.Error("expected an error because the bucket name is empty but got no error")
-	}
-
-	_, err = fetchObjectValues("myobject", "mybucket", nil)
-	if err == nil {
-		t.Error("expected an error because the cluster is nil but got no error")
-	}
-}
-
-func TestSaveObject(t *testing.T) {
-	cluster := &MockStorageCluster{
-		Running: true,
-	}
-
-	obj := riak.Object{}
-
-	err := saveObject(&obj, "bucket", cluster)
-	if err != nil {
-		t.Error("expected nil error got ", err)
-	}
-
-	err = saveObject(&obj, "", cluster)
-	if err == nil {
-		t.Error("expected an error due to empty bucket name but got a nil error")
-	}
-
-	err = saveObject(nil, "bucket", cluster)
-	if err == nil {
-		t.Error("expected an error because the obj is nil but go no error", err)
-	}
-
-	err = saveObject(&obj, "bucket", nil)
-	if err == nil {
-		t.Error("expected an error because the cluster is nil but go no error", err)
-	}
-}
-
-func TestDeleteObject(t *testing.T) {
-	cluster := &MockStorageCluster{
-		Running: true,
-	}
-
-	err := deleteObject("myobject", "bucket", cluster)
-	if err != nil {
-		t.Error("expected nil error got ", err)
-	}
-
-	err = deleteObject("", "bucket", cluster)
-	if err == nil {
-		t.Error("expected an empty key error but got nil error", err)
-	}
-
-	err = deleteObject("myobject", "", cluster)
-	if err == nil {
-		t.Error("expected an empty bucket error but got nil error", err)
-	}
-
-	err = deleteObject("myobject", "bucket", nil)
-	if err == nil {
-		t.Error("expected an nil cluster error but got nil error", err)
-	}
-}
-
 func TestValidateURIKeyset(t *testing.T) {
 	var keyset map[string]URISignerKeyset
 
@@ -276,43 +162,5 @@ func TestValidateURIKeyset(t *testing.T) {
 	// now validate it, expect an erro due to missing a matching key kid to renewal_kid
 	if err := validateURIKeyset(keyset); err == nil {
 		t.Errorf("validateURIKeyset(): expected an error")
-	}
-}
-
-func TestGetRiakCluster(t *testing.T) {
-	var cfg Config
-	mockDB, mock, err := sqlmock.New()
-	if err != nil {
-		t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
-	}
-	defer mockDB.Close()
-
-	db := sqlx.NewDb(mockDB, "sqlmock")
-	defer db.Close()
-
-	rows1 := sqlmock.NewRows([]string{"s.host_name", "s.domain_name"})
-	rows1.AddRow("www", "devnull.com")
-	mock.ExpectQuery("SELECT").WillReturnRows(rows1)
-
-	cfg.RiakAuthOptions = nil
-	if _, err := getRiakCluster(db, cfg); err == nil {
-		t.Errorf("expected an error due to nil RiakAuthoptions in the config but, go no error.")
-	}
-
-	cfg.RiakAuthOptions = &riak.AuthOptions{
-		User:      "riakuser",
-		Password:  "password",
-		TlsConfig: &tls.Config{},
-	}
-
-	if _, err := getRiakCluster(db, cfg); err != nil {
-		t.Errorf("expected no errors, actual: %s.", err)
-	}
-
-	rows2 := sqlmock.NewRows([]string{"s.host_name", "s.domain_name"})
-	mock.ExpectQuery("SELECT").WillReturnRows(rows2)
-
-	if _, err := getRiakCluster(db, cfg); err == nil {
-		t.Errorf("expected an error due to no available riak servers.")
 	}
 }
